+++ src/components/unisonInput/index.tsx
@@ -1,36 +1,22 @@
-import { useCallback } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-
 import styles from './UnisonInput.css';
 
-export const UnisonKnob: React.VFC = () => {
-  const dispatch = useDispatch();
+interface Props {
+  unison: number;
+  changeHandler: (e: React.ChangeEvent<HTMLInputElement>) => void;
+}
 
-  const unison = useSelector((state) => state.unison);
-
-  const changeHandler = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      dispatch({
-        type: 'updateUnison',
-        payload: { count: Number(e.target.value) },
-      });
-    },
-    [dispatch]
-  );
-
-  return (
-    <div>
-      <div className={styles.text}>
-        UNISON:&nbsp;
-        <input
-          className={styles.numberInput}
-          type="number"
-          value={unison}
-          min="1"
-          max="8"
-          onChange={changeHandler}
-        />
-      </div>
+export const UnisonInput: React.VFC<Props> = ({ unison, changeHandler }) => (
+  <div>
+    <div className={styles.text}>
+      UNISON:&nbsp;
+      <input
+        className={styles.numberInput}
+        type="number"
+        value={unison}
+        min="1"
+        max="8"
+        onChange={changeHandler}
+      />
     </div>
-  );
-};
+  </div>
+);
