+++ src/components/waveformSelectbox/index.tsx
@@ -1,38 +1,25 @@
-import { useCallback } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-
 import { Waveform, waveforms } from '../../domains/waveform';
-import { updateWaveform } from '../../store';
 
 import styles from './WaveformSelectbox.css';
 
-export const WaveformSelectbox: React.VFC = () => {
-  const dispatch = useDispatch();
+interface Props {
+  waveform: Waveform;
+  changeHandler: (e: React.ChangeEvent<HTMLSelectElement>) => void;
+}
 
-  const onChange = useCallback(
-    (e: React.ChangeEvent<HTMLSelectElement>) => {
-      dispatch(
-        updateWaveform({
-          waveform: e.target.value as Waveform,
-        })
-      );
-    },
-    [dispatch]
-  );
-
-  const currentWaveform = useSelector((state) => state.waveform);
-
-  return (
-    <select
-      className={styles.selectbox}
-      onChange={onChange}
-      value={currentWaveform}
-    >
-      {waveforms.map((waveform) => (
-        <option key={waveform} value={waveform}>
-          {waveform}
-        </option>
-      ))}
-    </select>
-  );
-};
+export const WaveformSelectbox: React.VFC<Props> = ({
+  waveform,
+  changeHandler,
+}) => (
+  <select
+    className={styles.selectbox}
+    onChange={changeHandler}
+    value={waveform}
+  >
+    {waveforms.map((waveform) => (
+      <option key={waveform} value={waveform}>
+        {waveform}
+      </option>
+    ))}
+  </select>
+);
